+++ lib/rails_audit_log/graphql/types/base_subscription.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module RailsAuditLog
+  module Graphql
+    module Types
+      class BaseSubscription < GraphQL::Schema::Subscription
+      end
+    end
+  end
+end