+++ spec/dummy/app/graphql/dummy_schema.rb
@@ -1,3 +1,5 @@
 class DummySchema < GraphQL::Schema
   query Types::QueryType
+  subscription Types::SubscriptionType
+  use GraphQL::Subscriptions, broadcast: true
 end